Performance: add ImageDPIRaw
The ImageDPI function currently creates an image, copies the C byte data
into it, and then frees the underlying data. This approach becomes
inefficient when used with glTexture due to the creation of a temporary
intermediate image, which causes a performance bottleneck.

To optimize, we can eliminate the data copying step by directly
returning the pointer. Profiling indicates that the copy and image
creation account for approximately 40% of the time spent when quickly
switching pages in an OpenGL viewer.